+++ src/sample/Controller/finalPath.java
@@ -1,0 +1,24 @@
+package sample.Controller;
+
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Node;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+import java.io.File;
+
+public class finalPath {
+    public String getfinalpath(String getfile) {
+        File file = new File(getfile);
+        String pathname = file.getAbsolutePath();
+
+        String out = "";
+        for (int i = 0 ; i < getfile.length() ; i++){
+            if (getfile.charAt(i) ==  '\\'){
+                out = out + "/";
+            }else out = out + getfile.charAt(i);
+        }
+        return  out;
+    }
+
+}
