+++ src/sample/listDictionary/textfield
@@ -8,6 +8,7 @@
  { dad ; bố } 
  { day ; ngày } 
  { demon ; quỷ } 
+ { dmmmm ; dm minh } 
  { down ; xuống } 
  { dungeon ; ngục } 
  { ear ; tai } 
@@ -44,5 +45,6 @@
  { interpol ; cảnh sát quốc tế } 
  { laugh ; ok pạn } 
  { nguyễn mạnh tiến ; hải phòng no hito =))) } 
+ { oop ; lap trinh huong doi tuong } 
  { sigma ; xyz } 
  { zeus ; name of a god }
